Got rid of some long standing conversion warning by providing the proper casts.
git-svn-id: http://yate.null.ro/svn/yate/trunk@3377 acf43c95-373e-0410-b603-e72c3f656dc1

+++ engine/DataFormat.cpp
@@ -511,9 +511,9 @@ bool DataConsumer::synchronize(DataSource* source)
     DDebug(DebugInfo,"Offsetting consumer %s timestamps by " FMT64 " [%p]",
 	(override ? "override" : "regular"),dt,this);
     if (override)
-	m_overrideTsDelta = dt;
+	m_overrideTsDelta = (long int)dt;
     else
-	m_regularTsDelta = dt;
+	m_regularTsDelta = (long int)dt;
     return true;
 }
